update v2
Mirjam's picture is not rendering because it wants an image file with a V2 suffix but there is only one without, so either the latest version of the file is not included or you need to change the file name.
At the end of people's profiles, there are sometimes commas, spaces or full stops. I think it is cleaner to always have a full stop at the end. [not really important, would merge without requesting this change if there were no other comments].
You need to rebuild the website locally (as per the README) and commit this rebuild. Otherwise when merging to master, the website will just reload the old version.